tools/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hokku-config"
version = "1.0.0"
description = "Configuration tool for Hokku e-paper frame"
requires-python = ">=3.8"
license = {text = "GPL-3.0"}
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
hokku-config = "hokku_config:main"

[tool.setuptools]
py-modules = ["hokku_config"]
